+++ CHANGES.md
@@ -1,5 +1,6 @@
 # Change Log
  - Oct. 2016
+   - Added `Read-Choice` module
    - Added `Write-HtmlNode` module
  - Aug. 2016
    - Added `Recycle` module